Tracer classes segregation
