+++ tracer/src/Datadog.Tracer.Native/Datadog.Tracer.Native.vcxproj
@@ -335,6 +335,10 @@
     <ClInclude Include="stats.h" />
     <ClInclude Include="Synchronized.hpp" />
     <ClInclude Include="threadUtils.h" />
+    <ClInclude Include="tracer_handler_module_method.h" />
+    <ClInclude Include="tracer_integration_definition.h" />
+    <ClInclude Include="tracer_method_rewriter.h" />
+    <ClInclude Include="tracer_rejit_preprocessor.h" />
     <ClInclude Include="tracer_tokens.h" />
   </ItemGroup>
   <ItemGroup>
@@ -385,6 +389,10 @@
     <ClCompile Include="rejit_preprocessor.cpp" />
     <ClCompile Include="rejit_work_offloader.cpp" />
     <ClCompile Include="threadUtils.cpp" />
+    <ClCompile Include="tracer_handler_module_method.cpp" />
+    <ClCompile Include="tracer_integration_definition.cpp" />
+    <ClCompile Include="tracer_method_rewriter.cpp" />
+    <ClCompile Include="tracer_rejit_preprocessor.cpp" />
     <ClCompile Include="tracer_tokens.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
