+++ tracer/src/Datadog.Tracer.Native/cor_profiler.h
@@ -14,6 +14,7 @@
 #include "il_rewriter.h"
 #include "integration.h"
 #include "rejit_preprocessor.h"
+#include "tracer_rejit_preprocessor.h"
 #include "debugger_rejit_preprocessor.h"
 #include "rejit_handler.h"
 #include <unordered_set>
@@ -45,7 +46,6 @@ class FaultTolerantRewriter;
 
 namespace trace
 {
-
 class CorProfiler : public CorProfilerBase
 {
 private:
@@ -207,7 +207,8 @@ public:
     //
     // Tracer Integration methods #2
     //
-    long RegisterCallTargetDefinitions(WCHAR* id, CallTargetDefinition2* items, int size, UINT32 enabledCategories = -1);
+    long RegisterCallTargetDefinitions(WCHAR* id, CallTargetDefinition2* items, int size,
+                                       UINT32 enabledCategories = -1);
     long EnableCallTargetDefinitions(UINT32 enabledCategories);
     long DisableCallTargetDefinitions(UINT32 disabledCategories);
