+++ tracer/src/Datadog.Tracer.Native/method_rewriter.h
@@ -36,28 +36,6 @@ public:
     virtual ~MethodRewriter() = default;
 };
 
-
-class TracerMethodRewriter : public MethodRewriter
-{
-private:
-    ILInstr* CreateFilterForException(ILRewriterWrapper* rewriter, mdTypeRef exception, mdTypeRef type_ref, mdMethodDef containsCallTargetBubbleUpException, ULONG exceptionValueIndex);
-
-public:
-
-    TracerMethodRewriter(CorProfiler* corProfiler) : MethodRewriter(corProfiler)
-    {
-    }
-
-    HRESULT Rewrite(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler, ICorProfilerFunctionControl* pFunctionControl) override;
-
-    InstrumentingProducts GetInstrumentingProduct(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler) override;
-    WSTRING GetInstrumentationId(RejitHandlerModule* moduleHandler, RejitHandlerModuleMethod* methodHandler) override;
-
-    std::tuple<WSTRING, WSTRING> GetResourceNameAndOperationName(const ComPtr<IMetaDataImport2>& metadataImport,
-                                                             const FunctionInfo* caller,
-                                                             TracerTokens* tracerTokens) const;
-};
-
 } // namespace trace
 
 #endif // DD_CLR_PROFILER_METHOD_REWRITER_H_
