+++ tracer/src/Datadog.Tracer.Native/rejit_handler.cpp
@@ -130,23 +130,6 @@ MethodRewriter* RejitHandlerModuleMethod::GetMethodRewriter()
 }
 
 //
-// TracerRejitHandlerModuleMethod
-//
-
-TracerRejitHandlerModuleMethod::TracerRejitHandlerModuleMethod(
-    mdMethodDef methodDef, RejitHandlerModule* module, const FunctionInfo& functionInfo,
-    const IntegrationDefinition& integrationDefinition, std::unique_ptr<MethodRewriter> methodRewriter) :
-    RejitHandlerModuleMethod(methodDef, module, functionInfo, std::move(methodRewriter)),
-    m_integrationDefinition(std::make_unique<IntegrationDefinition>(integrationDefinition))
-{
-}
-
-IntegrationDefinition* TracerRejitHandlerModuleMethod::GetIntegrationDefinition()
-{
-    return m_integrationDefinition.get();
-}
-
-//
 // RejitHandlerModule
 //
