+++ tracer/src/Datadog.Tracer.Native/tracer_rejit_preprocessor.h
@@ -1,0 +1,42 @@
+#pragma once
+
+#include "cor.h"
+#include "corprof.h"
+#include "integration.h"
+#include "module_metadata.h"
+#include "rejit_preprocessor.h"
+#include <future>
+
+
+namespace trace
+{
+class CorProfiler;
+class RejitHandler;
+class RejitWorkOffloader;
+class RejitHandlerModuleMethod;
+class RejitHandlerModule;
+struct FunctionInfo;
+
+/// <summary>
+/// TracerRejitPreprocessor
+/// </summary>
+class TracerRejitPreprocessor : public trace::RejitPreprocessor<IntegrationDefinition>
+{
+public:
+    using RejitPreprocessor::RejitPreprocessor;
+
+protected:
+    const MethodReference& GetTargetMethod(const IntegrationDefinition& integrationDefinition) final;
+    const bool GetIsDerived(const IntegrationDefinition& definition) final;
+    const bool GetIsInterface(const IntegrationDefinition& definition) final;
+    const bool GetIsExactSignatureMatch(const IntegrationDefinition& definition) final;
+    const bool GetIsEnabled(const IntegrationDefinition& definition) final;
+    const bool SupportsSelectiveEnablement() final;
+
+    const std::unique_ptr<RejitHandlerModuleMethod>
+    CreateMethod(mdMethodDef methodDef, RejitHandlerModule* module, const FunctionInfo& functionInfo,
+                 const IntegrationDefinition& integrationDefinition) final;
+    bool ShouldSkipModule(const ModuleInfo& moduleInfo, const IntegrationDefinition& integrationDefinition) final;
+};
+
+} // namespace trace
